Begin work on moving Stoner.Plot* modules to a Stoner.plot sub-package
